1.1.1 branch

+++ Decorator/src/test/java/companyB/decorated/test/FileLoadedTestClassTest.java
@@ -86,10 +86,7 @@ public class FileLoadedTestClassTest
         }
         finally
         {
-            if(file.exists())
-            {
-                file.deleteOnExit();
-            }
+            if(file.exists()) file.deleteOnExit();
         }
 
     }
